chore(piercing) PR review changes for news tab
