+++ productivity-suite/app/fragments/news/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "solid-hackernews2",
+  "name": "productivity-suite-news-fragment",
   "scripts": {
     "dev": "solid-start dev",
     "build": "solid-start build",
-    "start": "npx wrangler dev --local --env local"
+    "start": "npx wrangler dev --local"
   },
   "type": "module",
   "main": "./dist/index.js",
@@ -12,7 +12,8 @@
     "solid-start-node": "^0.2.0",
     "typescript": "^4.8.4",
     "vite": "^3.1.8",
-    "wrangler": "^2.1.13"
+    "wrangler": "^2.1.13",
+		"shared": "file:../../shared"
   },
   "dependencies": {
     "@solidjs/meta": "^0.28.0",
