+++ productivity-suite/app/fragments/todos/package.json
@@ -28,6 +28,7 @@
 		"@vitejs/plugin-react": "^2.0.1",
 		"typescript": "^4.6.4",
 		"vite": "^3.0.7",
-		"wrangler": "^2.1.9"
+		"wrangler": "^2.1.9",
+		"shared": "file:../../shared"
 	}
 }
