+++ productivity-suite/app/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./api";
+export * from "./addDefaultFnPlugin";
 
 export type { Todo, TodoList } from "./todos";
